+++ object/ObjString.h
@@ -2,7 +2,7 @@
  * @Author: nestcc 
  * @Date: 2021/4/4 22:48
  * @LastEditors: Nestcc
- * @LastEditTime: 2021-04-22 14:10:40
+ * @LastEditTime: 2021-04-28 10:40:15
  * @Discription: 
  */
 
@@ -20,7 +20,7 @@ public:
 
     ObjString() = delete;
     ObjString(VM *vm, const std::string &str);
-    ObjString(VM *vm, const char *str, uint32_t ssize);
+    ObjString(VM *vm, const char *str, uint64_t ssize);
     
     uint64_t hash_value() const override;
